llamaindex/rag/app/requirements.txt
llama-index-storage-docstore-redis==0.3.0
llama-index-vector-stores-redis==0.4.0
llama-index-embeddings-huggingface==0.5.0
llama-index-readers-file==0.4.3
llama-index-llms-ollama==0.5.0
llama-index-core>=0.12.4
fastapi==0.115.6
uvicorn==0.34.0
pandas==2.2.3 
torch==2.5.1
pydantic